+++ src/components/Menu.tsx
@@ -31,6 +31,7 @@ import LockOpenIcon from '@material-ui/icons/LockOpen'
 import HomeIcon from '@material-ui/icons/Home'
 import DonutLargeIcon from '@material-ui/icons/DonutLarge'
 import MailIcon from '@material-ui/icons/Mail'
+import { Cluster as ClusterModel } from '@redkubes/otomi-api-client-axios'
 import { useApi } from '../hooks/api'
 import { mainStyles } from '../theme'
 import snack from '../utils/snack'
@@ -68,6 +69,7 @@ interface Props {
 export default ({ teamId }: Props): React.ReactElement => {
   const { pathname } = useLocation()
   const {
+    cluster,
     dirty,
     mode,
     user: { isAdmin },
@@ -165,6 +167,8 @@ export default ({ teamId }: Props): React.ReactElement => {
       <Collapse component='li' in={collapseSettings} timeout='auto' unmountOnExit>
         <List className={classes.settingsList} disablePadding>
           {Object.keys(settingIds).map((id) => {
+            // TODO: fix this hack with a generic x-provider approach?
+            if (cluster.provider !== ClusterModel.ProviderEnum.azure && id === 'azure') return undefined
             return (
               <StyledMenuItem
                 key={id}